backend/requirements.txt
amqp==5.3.1
beautifulsoup4==4.13.4
billiard==4.2.1
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
contourpy==1.3.3
cryptography==46.0.1
cssselect2==0.8.0
cycler==0.12.1
dnspython==2.8.0
et_xmlfile==2.0.0
eventlet==0.40.2
filelock==3.19.1
Flask==3.1.1
flask-cors==6.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Moment==1.0.6
Flask-SQLAlchemy==3.1.1
flower==2.0.1
fonttools==4.59.2
fsspec==2025.9.0
fuzzywuzzy==0.18.0
gensim==4.3.3
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
gunicorn==21.2.0
hf-xet==1.1.10
honcho==1.1.0
httplib2==0.31.0
huggingface-hub==0.34.4
humanize==4.13.0
idna==3.10
imbalanced-learn==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
kombu==5.5.4
Levenshtein==0.27.1
lxml==6.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.5
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
parrot @ git+https://github.com/PrithivirajDamodaran/Parrot_Paraphraser.git@03084c54b64019ba5fa0b620b9c70ad81123e458
pillow==11.3.0
prometheus_client==0.22.1
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.32.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydyf==0.11.0
PyJWT==2.10.1
pymongo==4.13.2
pyparsing==3.2.4
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-Levenshtein==0.27.1
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.14.1
redis==6.2.0
regex==2025.9.1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
sentence-transformers==5.0.0
sentencepiece==0.2.1
setuptools==80.9.0
six==1.17.0
smart_open==7.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sympy==1.14.0
textblob==0.19.0
threadpoolctl==3.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.21.4
torch==2.7.1
tornado==6.5.2
tqdm==4.67.1
transformers==4.53.2
triton==3.3.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==66.0
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.3
xgboost==3.0.5
zopfli==0.2.3.post1
